Fix case where string encoding overhead is larger than needed
According to the spec "serializers SHOULD use the format which
represents the data in the smallest number of bytes". This is not true
for strings between 32 and (2^8)-1 bytes which has an encoding
overhead of 3 bytes instead of the required 2.

My guess is that the somewhat strange conditional in fallback.py:598
is to make sure that old raw type test cases in test_case.py that test
for encoding overhead succeeds.